# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a water station.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://poseidon-backend.fly.dev/",
    "basePath": "/",
    "paths": {
...
        "/likes": {
            "get": {
                "description": "Get all likes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Show all likes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Like"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing like",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Update a like",
                "parameters": [
                    {
                        "description": "Like",
                        "name": "like",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Like"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Like"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new like",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Create a like",
                "parameters": [
                    {
                        "description": "Like",
                        "name": "like",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Like"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Like"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing like",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Delete a like",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Like ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/refill_stations": {
            "get": {
                "description": "Get all refill stations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_stations"
                ],
                "summary": "Show all refill stations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.RefillStation"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing refill station",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_stations"
                ],
                "summary": "Update a refill station",
                "parameters": [
                    {
                        "description": "Refill Station",
                        "name": "station",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.RefillStation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.RefillStation"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new refill station",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_stations"
                ],
                "summary": "Create a refill station",
                "parameters": [
                    {
                        "description": "Refill Station",
                        "name": "station",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.RefillStation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.RefillStation"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing refill station",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_stations"
                ],
                "summary": "Delete a refill station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Refill Station ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/refill_stations/{id}": {
            "get": {
                "description": "Get a refill station by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_stations"
                ],
                "summary": "Get a refill station by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Refill Station ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.RefillStation"
                        }
                    }
                }
            }
        },
        "/refill_stations/{id}/reviews": {
            "get": {
                "description": "Get the average review score for a refill station by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_station_reviews"
                ],
                "summary": "Get the average review score for a refill station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Refill Station ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "number"
                        }
                    }
                }
            }
        },
        "/refillstation_like": {
            "get": {
                "description": "Check if a specific user likes a specific refill station",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Check if a user likes a refill station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Refill Station ID",
                        "name": "refillstationId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.Like": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "station_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "database.RefillStation": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "offered_water_types": {
                    "type": "string"
                },
                "opening_times": {
                    "type": "string"
                },
                "refill_station_image": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "type": "string"
                },
                "water_source": {
                    "type": "string"
                }
            }
        },
...
    }
}